gemini_api/requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
Flask==3.1.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
Pillow==10.4.0
pluggy==1.5.0
pytest==8.3.4
pytest-flask==1.3.0
python-dotenv==1.0.1
requests==2.32.3
tomli==2.2.1
urllib3==2.2.3
Werkzeug
zipp==3.21.0